+++ apps/scrybe/core/database/src/main/kotlin/dev/scrybe/core/database/di/DatabaseModule.kt
@@ -11,6 +11,7 @@ import dev.scrybe.core.database.AppDatabase
 import dev.scrybe.core.database.MIGRATION_10_11
 import dev.scrybe.core.database.MIGRATION_11_12
 import dev.scrybe.core.database.MIGRATION_12_13
+import dev.scrybe.core.database.MIGRATION_13_14
 import dev.scrybe.core.database.MIGRATION_4_5
 import dev.scrybe.core.database.MIGRATION_5_6
 import dev.scrybe.core.database.MIGRATION_6_7
@@ -42,6 +43,7 @@ object DatabaseModule {
                 MIGRATION_10_11,
                 MIGRATION_11_12,
                 MIGRATION_12_13,
+                MIGRATION_13_14,
             ).fallbackToDestructiveMigrationFrom(1, 2, 3)
             .build()
 
@@ -74,4 +76,7 @@ object DatabaseModule {
 
     @Provides
     fun providesSessionTaskDao(database: AppDatabase) = database.sessionTaskDao()
+
+    @Provides
+    fun providesCustomRecordingTypeDao(database: AppDatabase) = database.customRecordingTypeDao()
 }